Move service REST API code to dedicated module
Update the service and scabbard REST API modules to isolate the REST API
code into submodules.
