+++ libsplinter/src/service/mod.rs
@@ -39,12 +39,13 @@ pub mod error;
 mod factory;
 mod processor;
 mod registry;
+mod rest_api;
 pub mod scabbard;
 mod sender;
 
 use std::any::Any;
 
-pub use factory::{ServiceEndpoint, ServiceFactory};
+pub use factory::ServiceFactory;
 pub use processor::JoinHandles;
 pub use processor::ServiceProcessor;
 pub use processor::ShutdownHandle;
