+++ libsplinter/src/service/rest_api.rs
@@ -1,0 +1,40 @@
+// Copyright 2019 Cargill Incorporated
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::sync::Arc;
+
+use crate::actix_web::{web, Error as ActixError, HttpRequest, HttpResponse};
+use crate::futures::Future;
+use crate::rest_api::Method;
+
+use super::Service;
+
+type Handler = Arc<
+    dyn Fn(
+            HttpRequest,
+            web::Payload,
+            &dyn Service,
+        ) -> Box<dyn Future<Item = HttpResponse, Error = ActixError>>
+        + Send
+        + Sync
+        + 'static,
+>;
+
+#[derive(Clone)]
+pub struct ServiceEndpoint {
+    pub service_type: String,
+    pub route: String,
+    pub method: Method,
+    pub handler: Handler,
+}
