+++ libsplinter/src/service/scabbard/rest_api.rs
@@ -1,0 +1,187 @@
+// Copyright 2019 Cargill Incorporated
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::collections::HashMap;
+use std::sync::Arc;
+
+use transact::protocol::batch::BatchPair;
+use transact::protos::FromBytes;
+
+use crate::actix_web::{web, Error as ActixError, HttpResponse};
+use crate::futures::{stream::Stream, Future, IntoFuture};
+use crate::rest_api::{new_websocket_event_sender, EventSender, Method, Request};
+use crate::service::rest_api::ServiceEndpoint;
+
+use super::error::StateSubscriberError;
+use super::state::{StateChangeEvent, StateSubscriber};
+use super::{Scabbard, SERVICE_TYPE};
+
+struct WsStateSubscriber {
+    sender: EventSender<StateChangeEvent>,
+}
+
+impl StateSubscriber for WsStateSubscriber {
+    fn handle_event(&self, event: StateChangeEvent) -> Result<(), StateSubscriberError> {
+        self.sender.send(event).map_err(|_| {
+            debug!(
+                "Dropping scabbard state change event and unsubscribing due to websocket being
+                 closed"
+            );
+            StateSubscriberError::Unsubscribe
+        })
+    }
+}
+
+pub fn make_subscribe_endpoint() -> ServiceEndpoint {
+    ServiceEndpoint {
+        service_type: SERVICE_TYPE.into(),
+        route: "/ws/subscribe".into(),
+        method: Method::Get,
+        handler: Arc::new(move |request, payload, service| {
+            let scabbard = match service.as_any().downcast_ref::<Scabbard>() {
+                Some(s) => s,
+                None => {
+                    return Box::new(HttpResponse::InternalServerError().finish().into_future())
+                }
+            };
+
+            let mut query =
+                match web::Query::<HashMap<String, String>>::from_query(request.query_string()) {
+                    Ok(query) => query,
+                    Err(_) => return Box::new(HttpResponse::BadRequest().finish().into_future()),
+                };
+
+            let last_seen_event_id = query.remove("last_seen_event");
+
+            if let Some(ref id) = last_seen_event_id {
+                debug!("Getting all state-delta events since {}", id);
+            } else {
+                debug!("Getting all state-delta events");
+            }
+
+            let unseen_events = match scabbard.get_events_since(last_seen_event_id) {
+                Ok(events) => events,
+                Err(err) => {
+                    error!("Unable to load unseen scabbard events: {}", err);
+                    return Box::new(HttpResponse::InternalServerError().finish().into_future());
+                }
+            };
+
+            let request = Request::from((request, payload));
+            match new_websocket_event_sender(request, Box::new(unseen_events)) {
+                Ok((sender, res)) => {
+                    if let Err(err) =
+                        scabbard.add_state_subscriber(Box::new(WsStateSubscriber { sender }))
+                    {
+                        error!("Unable to add scabbard event sender: {}", err);
+                        return Box::new(
+                            HttpResponse::InternalServerError().finish().into_future(),
+                        );
+                    }
+                    Box::new(res.into_future())
+                }
+                Err(err) => {
+                    debug!("Failed to create websocket: {:?}", err);
+                    Box::new(HttpResponse::InternalServerError().finish().into_future())
+                }
+            }
+        }),
+    }
+}
+
+pub fn make_add_batches_to_queue_endpoint() -> ServiceEndpoint {
+    ServiceEndpoint {
+        service_type: SERVICE_TYPE.into(),
+        route: "/batches".into(),
+        method: Method::Post,
+        handler: Arc::new(move |_, payload, service| {
+            let scabbard = match service.as_any().downcast_ref::<Scabbard>() {
+                Some(s) => s,
+                None => {
+                    return Box::new(HttpResponse::InternalServerError().finish().into_future())
+                }
+            }
+            .clone();
+
+            Box::new(
+                payload
+                    .from_err::<ActixError>()
+                    .fold(web::BytesMut::new(), move |mut body, chunk| {
+                        body.extend_from_slice(&chunk);
+                        Ok::<_, ActixError>(body)
+                    })
+                    .into_future()
+                    .and_then(move |body| {
+                        let batches: Vec<BatchPair> = match Vec::from_bytes(&body) {
+                            Ok(b) => b,
+                            Err(_) => return HttpResponse::BadRequest().finish().into_future(),
+                        };
+
+                        match scabbard.add_batches(batches) {
+                            Ok(Some(link)) => HttpResponse::Accepted().json(link).into_future(),
+                            Ok(None) => HttpResponse::BadRequest().finish().into_future(),
+                            Err(_) => HttpResponse::InternalServerError().finish().into_future(),
+                        }
+                    }),
+            )
+        }),
+    }
+}
+
+pub fn make_get_batch_status_endpoint() -> ServiceEndpoint {
+    ServiceEndpoint {
+        service_type: SERVICE_TYPE.into(),
+        route: "/batch_statuses".into(),
+        method: Method::Get,
+        handler: Arc::new(move |req, _, service| {
+            let scabbard = match service.as_any().downcast_ref::<Scabbard>() {
+                Some(s) => s,
+                None => {
+                    return Box::new(HttpResponse::InternalServerError().finish().into_future())
+                }
+            }
+            .clone();
+            let query: web::Query<HashMap<String, String>> =
+                if let Ok(q) = web::Query::from_query(req.query_string()) {
+                    q
+                } else {
+                    return Box::new(
+                        HttpResponse::BadRequest()
+                            .json(json!({
+                                "message": "Invalid query"
+                            }))
+                            .into_future(),
+                    );
+                };
+
+            let ids = if let Some(ids) = query.get("ids") {
+                ids.split(',').map(String::from).collect()
+            } else {
+                return Box::new(
+                    HttpResponse::BadRequest()
+                        .json(json!({
+                            "message": "No batch IDs specified"
+                        }))
+                        .into_future(),
+                );
+            };
+
+            if let Ok(status) = scabbard.get_batch_info(ids) {
+                Box::new(HttpResponse::Ok().json(status).into_future())
+            } else {
+                Box::new(HttpResponse::InternalServerError().finish().into_future())
+            }
+        }),
+    }
+}
